Add perfect grid sizing functionality.
This is an experimental version to make a perfect grid (no gaps). It's
hard coded and not suitable for generic use yet.

For instance, it was added a 'sizes' option that is used to randomily
pick a size for each element on grid. The plugin might disrespect these
sizes for some elements in order to get a perfect grid.